incorporates changes requested in #41 for minutes item component
