--- src/components/Cards/MinutesItemCard/index.ts
@@ -1,1 +1,0 @@
-export { default as MinutesItemCard } from "./MinutesItemCard";
