+++ src/components/Details/MinutesItem/index.ts
@@ -1,0 +1,1 @@
+export { default as MinutesItem } from "./MinutesItem";
